refactor: js utils
