--- repos/kaleido/js/src/plotly/is-non-empty-string.js
@@ -1,5 +1,0 @@
-function isNonEmptyString (v) {
-  return typeof v === 'string' && v.length > 0
-}
-
-module.exports = isNonEmptyString
