--- repos/kaleido/js/src/plotly/is-positive-numeric.js
@@ -1,7 +1,0 @@
-const isNumeric = require('fast-isnumeric')
-
-function isPositiveNumeric (v) {
-  return isNumeric(v) && v > 0
-}
-
-module.exports = isPositiveNumeric
