+++ repos/kaleido/js/src/utils/has-propperties-of-object.js
