+++ repos/kaleido/js/src/utils/is-non-empty-string.js
